% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{missInfer}
\alias{missInfer}
\title{Estimate coefficients under complex missing mechanisms}
\usage{
missInfer(
  X,
  Z = NULL,
  y,
  interest_var = 1:NCOL(X),
  intercept = T,
  thres_w = 0.01,
  pi.method = c("kernel", "glmnet"),
  dim.reduction = c("separate", "onlyQ")
)
}
\arguments{
\item{X}{input data matrix.}

\item{Z}{surrogate outcome(s). If a surrogate outcome is not available, then Z can be NULL.}

\item{y}{response variable.}

\item{interest_var}{variables to be estimated. Default is all the variables of X.}

\item{intercept}{intercept (default=TRUE).}

\item{thres_w}{threshold to avoid too small inverse pi value.}

\item{pi.method}{method for estimating the inverse pi function. The kernel method is default but the glmnet method is also available. Should be either "kernel" or "glmnet".}

\item{dim.reduction}{models for estimating a reduced dimension. The imputation model is default but both nuisance models are used for the dimension reduction. Should be either "separate" or "onlyQ".}
}
\value{
a list that contains the initial coefficients, debaised coefficients, dimensionality of reduced dimension
}
\description{
Estimate coefficients under complex missing mechanisms
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
