% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{missInfer}
\alias{missInfer}
\title{Estimate coefficients under complex missing mechanisms}
\usage{
missInfer(
  X,
  Z = NULL,
  y,
  interest_var = 1:4,
  intercept = T,
  thres_w = 0.01,
  pi.method = c("kernel", "glmnet"),
  dim.reduction = c("separate", "onlyQ")
)
}
\arguments{
\item{X}{Input data matrix}

\item{Z}{Surrogate outcome(s) Z can be a vector or a matrix (default: NULL)}

\item{y}{Response variable}

\item{interest_var}{Variables to be debiased (default: the first four variables)}

\item{intercept}{Intercept (default=TRUE)}

\item{thres_w}{Threshold to avoid too small inverse pi value (default: 0.01)}

\item{pi.method}{Method for estimating the inverse pi function. The kernel method is default but the glmnet method is also available. Should be either "kernel" or "glmnet".}

\item{dim.reduction}{Models for estimating a reduced dimension. The imputation model is default but both nuisance models are used for the dimension reduction. Should be either "separate" or "onlyQ".}
}
\value{
a list that contains the initial coefficients, debaised coefficients, dimensionality of reduced dimension
}
\description{
Estimate coefficients under complex missing mechanisms
}
\examples{
# Sample data generation
set.seed(1823)
p <- 8; interest_lst <- 1:8; y_type <- "binary"; alpha <- -1;
thres_w <- 0.01; missing.rate <- 0.5
samples <- sample.generation(p = p, n= n, y_type = y_type, alpha = alpha, missing.rate =missing.rate)
X <- samples$X; Z <- samples$Z; y <- samples$y # R <- samples$R;

# Obtain debiased estimators
missInfer(X = X, Z = Z, y = y, interest_var = 1:p, intercept = intercept, thres_w =  thres_w, pi.method = "kernel", dim.reduction = "onlyQ")
}
